refactor(orchestrator): extract usage, probe, and prompts into mixin modules
