+++ roboco/runtime/orchestrator.py
@@ -24,7 +24,6 @@ import httpx
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Coroutine
-    from uuid import UUID
 
     from roboco.services.llm import AgentRoute
     from roboco.services.task import TaskService
@@ -100,7 +99,10 @@ from roboco.runtime._helpers import (  # noqa: E402, F401
     _SlaBreach,
     get_agent_image,
 )
+from roboco.runtime._probe import AgentRateLimitMixin  # noqa: E402
+from roboco.runtime._prompts import AgentPromptsMixin  # noqa: E402
 from roboco.runtime._settings import AgentSettingsMixin  # noqa: E402
+from roboco.runtime._usage import AgentUsageMixin  # noqa: E402
 
 
 class AgentReadinessError(Exception):
@@ -111,7 +113,12 @@ class AgentReadinessError(Exception):
     """
 
 
-class AgentOrchestrator(AgentSettingsMixin):
+class AgentOrchestrator(
+    AgentSettingsMixin,
+    AgentUsageMixin,
+    AgentRateLimitMixin,
+    AgentPromptsMixin,
+):
     """
     Manages Claude Code containers for all agents.
 
@@ -2382,558 +2389,6 @@ class AgentOrchestrator(AgentSettingsMixin):
     # TOKEN USAGE INSTRUMENTATION
     # =========================================================================
 
-    async def _record_spawn_session(
-        self,
-        config: "OrchestratorAgentConfig",
-        task_id: str | None,
-    ) -> "UUID | None":
-        """Insert a row into agent_spawn_sessions after a successful spawn.
-
-        Returns the UUID of the created row so the caller can store it on
-        the AgentInstance for later direct-by-id lookup in
-        _finalize_spawn_session.  Returns None when the insert fails; a
-        missing session row must never block the spawn path.
-        """
-        try:
-            from uuid import uuid4 as _uuid4
-
-            from roboco.db.base import get_session_factory
-            from roboco.db.tables import AgentSpawnSessionTable
-
-            agent_slug = config.agent_id
-            team = get_agent_team(agent_slug) or "backend"
-            role = get_agent_role(agent_slug) or "developer"
-
-            session_id = _uuid4()
-            session_factory = get_session_factory()
-            async with session_factory() as db:
-                row = AgentSpawnSessionTable(
-                    id=session_id,
-                    agent_slug=agent_slug,
-                    team=team,
-                    role=role,
-                    model=config.model or "unknown",
-                    task_id=task_id,
-                    started_at=datetime.now(UTC),
-                )
-                db.add(row)
-                await db.commit()
-                logger.debug(
-                    "Spawn session recorded",
-                    agent_slug=agent_slug,
-                    session_id=str(session_id),
-                    task_id=task_id,
-                )
-            return session_id
-        except Exception as exc:
-            logger.warning(
-                "Failed to record spawn session",
-                agent_slug=config.agent_id,
-                error=str(exc),
-            )
-            return None
-
-    def _claude_session_id_for(self, agent_id: str) -> str | None:
-        """The orchestrator-assigned Claude session id for a running agent."""
-        instance = self._instances.get(agent_id)
-        return (
-            instance.config.claude_session_id if instance and instance.config else None
-        )
-
-    @staticmethod
-    def _usage_from_transcript(
-        agent_id: str, claude_session_id: str | None = None
-    ) -> tuple[int, int, int, int]:
-        """Sum token usage from the agent's Claude Code transcript.
-
-        The host ``~/.claude`` is mounted into the orchestrator, so transcripts
-        are readable here under ``projects/<cwd-dir>/<session-id>.jsonl``. When
-        the orchestrator-assigned ``claude_session_id`` is known we locate the
-        exact transcript by id across ANY project dir — review/coordinate roles
-        run at cwd ``/app`` so theirs lands in ``projects/-app``, not in a
-        per-agent ``projects/*-{slug}`` dir. Without an id we fall back to the
-        newest transcript in the agent's own workspace dir. Durable fallback for
-        the live SDK ``/usage/status`` fetch, which misses for short-lived or
-        torn-down agents. Returns zeros when no transcript is found.
-        """
-        from roboco.agent_sdk.transcript_usage import sum_transcript_usage
-
-        projects = Path.home() / ".claude" / "projects"
-        try:
-            if claude_session_id:
-                by_id = list(projects.glob(f"*/{claude_session_id}.jsonl"))
-                if by_id:
-                    return sum_transcript_usage(by_id[0])
-            jsonl = [
-                f
-                for d in projects.glob(f"*-{agent_id}")
-                if d.is_dir()
-                for f in d.glob("*.jsonl")
-            ]
-            if not jsonl:
-                return (0, 0, 0, 0)
-            newest = max(jsonl, key=lambda f: f.stat().st_mtime)
-            return sum_transcript_usage(newest)
-        except OSError:
-            return (0, 0, 0, 0)
-
-    async def _resolve_final_token_usage(
-        self, agent_id: str
-    ) -> tuple[int, int, int, int]:
-        """Resolve final token counts for a stopping agent.
-
-        Tries the live SDK ``/usage/status`` first; if that misses — the SDK's
-        in-memory counts race container teardown for short-lived agents — it
-        falls back to the agent's Claude Code transcript, which is durable and
-        mounted into this container. Returns
-        ``(input, output, cache_read, cache_write)``.
-        """
-        tokens = (0, 0, 0, 0)
-        sdk_url = f"http://roboco-agent-{agent_id}:{SDK_PORT}/usage/status"
-        try:
-            async with httpx.AsyncClient(timeout=3.0) as client:
-                resp = await client.get(sdk_url)
-                if resp.status_code == http_status.HTTP_200_OK:
-                    data = resp.json()
-                    tokens = (
-                        data.get("tokens_input", 0),
-                        data.get("tokens_output", 0),
-                        data.get("tokens_cache_read", 0),
-                        data.get("tokens_cache_write", 0),
-                    )
-        except Exception as sdk_exc:
-            logger.debug(
-                "Could not fetch final token counts from SDK",
-                agent_id=agent_id,
-                error=str(sdk_exc),
-            )
-
-        if not tokens[0] and not tokens[1]:
-            tin, tout, cr, cw = self._usage_from_transcript(
-                agent_id, self._claude_session_id_for(agent_id)
-            )
-            if tin or tout:
-                tokens = (tin, tout, cr, cw)
-        return tokens
-
-    async def _finalize_spawn_session(
-        self,
-        agent_id: str,
-        exit_reason: str = "stopped",
-    ) -> None:
-        """Close the open agent_spawn_sessions row for this agent.
-
-        Resolves final token counts (live SDK, with a durable transcript
-        fallback), calculates cost via the pricing module, then updates the DB
-        row with ended_at, token totals, exit_reason, and estimated_cost_usd.
-        Errors are caught and logged — finalization must never block stop_agent.
-        """
-        try:
-            from roboco.billing.pricing import calculate_cost
-            from roboco.db.base import get_session_factory
-            from roboco.db.tables import AgentSpawnSessionTable
-
-            # Resolve final token counts (live SDK, with transcript fallback).
-            (
-                tokens_input,
-                tokens_output,
-                tokens_cache_read,
-                tokens_cache_write,
-            ) = await self._resolve_final_token_usage(agent_id)
-
-            # Look up the model and usage_session_id from the running instance config.
-            model = "unknown"
-            instance = self._instances.get(agent_id)
-            if instance and instance.config:
-                model = instance.config.model or "unknown"
-            usage_session_id = instance.usage_session_id if instance else None
-
-            cost = calculate_cost(
-                model=model,
-                tokens_input=tokens_input,
-                tokens_output=tokens_output,
-                tokens_cache_read=tokens_cache_read,
-                tokens_cache_write=tokens_cache_write,
-            )
-
-            session_factory = get_session_factory()
-            async with session_factory() as db:
-                from sqlalchemy import select, update
-
-                # Prefer a direct lookup by the session UUID captured at spawn
-                # time; fall back to the (agent_slug, ended_at IS NULL) query
-                # for instances that pre-date the usage_session_id field.
-                if usage_session_id is not None:
-                    result = await db.execute(
-                        select(AgentSpawnSessionTable).where(
-                            AgentSpawnSessionTable.id == usage_session_id
-                        )
-                    )
-                else:
-                    result = await db.execute(
-                        select(AgentSpawnSessionTable)
-                        .where(
-                            AgentSpawnSessionTable.agent_slug == agent_id,
-                            AgentSpawnSessionTable.ended_at.is_(None),
-                        )
-                        .order_by(AgentSpawnSessionTable.started_at.desc())
-                        .limit(1)
-                    )
-                session_row = result.scalar_one_or_none()
-                if session_row is not None:
-                    await db.execute(
-                        update(AgentSpawnSessionTable)
-                        .where(AgentSpawnSessionTable.id == session_row.id)
-                        .values(
-                            ended_at=datetime.now(UTC),
-                            tokens_input=tokens_input,
-                            tokens_output=tokens_output,
-                            tokens_cache_read=tokens_cache_read,
-                            tokens_cache_write=tokens_cache_write,
-                            exit_reason=exit_reason,
-                            estimated_cost_usd=cost,
-                        )
-                    )
-                    await db.commit()
-                    logger.debug(
-                        "Spawn session finalized",
-                        agent_id=agent_id,
-                        session_id=str(session_row.id),
-                        tokens_input=tokens_input,
-                        tokens_output=tokens_output,
-                        estimated_cost_usd=cost,
-                    )
-        except Exception as exc:
-            logger.warning(
-                "Failed to finalize spawn session",
-                agent_id=agent_id,
-                error=str(exc),
-            )
-
-    @staticmethod
-    async def _fetch_agent_tokens(
-        client: httpx.AsyncClient, agent_id: str
-    ) -> tuple[int, int, int, int] | None:
-        """Fetch cumulative token counts from an agent's SDK usage endpoint.
-
-        Returns ``(input, output, cache_read, cache_write)`` or ``None`` when the
-        agent returns a non-200 status or has not accrued any tokens yet.
-        """
-        sdk_url = f"http://roboco-agent-{agent_id}:{SDK_PORT}/usage/status"
-        resp = await client.get(sdk_url)
-        if resp.status_code != http_status.HTTP_200_OK:
-            return None
-        data = resp.json()
-        tokens = (
-            data.get("tokens_input", 0),
-            data.get("tokens_output", 0),
-            data.get("tokens_cache_read", 0),
-            data.get("tokens_cache_write", 0),
-        )
-        if sum(tokens) == 0:
-            return None
-        return tokens
-
-    async def _resolve_active_tokens(
-        self, client: httpx.AsyncClient, agent_id: str
-    ) -> tuple[int, int, int, int] | None:
-        """Resolve live token counts for an active agent.
-
-        Tries the agent SDK's ``/usage/status`` first; on a zero/miss falls
-        back to the durable transcript (the SDK can report zero mid-run, the
-        same race the finalize path handles). Returns ``None`` when neither
-        source has any usage yet.
-        """
-        tokens = await self._fetch_agent_tokens(client, agent_id)
-        if tokens is not None:
-            return tokens
-        transcript = self._usage_from_transcript(
-            agent_id, self._claude_session_id_for(agent_id)
-        )
-        return transcript if any(transcript) else None
-
-    @staticmethod
-    async def _persist_token_snapshot(
-        session_factory: Any,
-        agent_id: str,
-        instance: AgentInstance,
-        tokens: tuple[int, int, int, int],
-    ) -> bool:
-        """Insert a token_usage_snapshots row and refresh the open session totals.
-
-        Returns True when a snapshot was written; False when the agent has no
-        open spawn-session row to attach it to.
-        """
-        from uuid import uuid4
-
-        from sqlalchemy import select, update
-
-        from roboco.db.tables import AgentSpawnSessionTable, TokenUsageSnapshotTable
-
-        tokens_input, tokens_output, tokens_cache_read, tokens_cache_write = tokens
-        async with session_factory() as db:
-            # Prefer a direct lookup by the session UUID captured at spawn time;
-            # fall back to the agent_slug heuristic for instances that pre-date
-            # the usage_session_id field.
-            if instance.usage_session_id is not None:
-                result = await db.execute(
-                    select(AgentSpawnSessionTable).where(
-                        AgentSpawnSessionTable.id == instance.usage_session_id
-                    )
-                )
-            else:
-                result = await db.execute(
-                    select(AgentSpawnSessionTable)
-                    .where(
-                        AgentSpawnSessionTable.agent_slug == agent_id,
-                        AgentSpawnSessionTable.ended_at.is_(None),
-                    )
-                    .order_by(AgentSpawnSessionTable.started_at.desc())
-                    .limit(1)
-                )
-            session_row = result.scalar_one_or_none()
-            if session_row is None:
-                return False
-
-            db.add(
-                TokenUsageSnapshotTable(
-                    id=uuid4(),
-                    agent_spawn_session_id=session_row.id,
-                    snapshotted_at=datetime.now(UTC),
-                    tokens_input=tokens_input,
-                    tokens_output=tokens_output,
-                    tokens_cache_read=tokens_cache_read,
-                    tokens_cache_write=tokens_cache_write,
-                )
-            )
-            await db.execute(
-                update(AgentSpawnSessionTable)
-                .where(AgentSpawnSessionTable.id == session_row.id)
-                .values(
-                    tokens_input=tokens_input,
-                    tokens_output=tokens_output,
-                    tokens_cache_read=tokens_cache_read,
-                    tokens_cache_write=tokens_cache_write,
-                )
-            )
-            await db.commit()
-            return True
-
-    async def _sweep_token_snapshots(self) -> None:
-        """Write a token_usage_snapshots row for each active agent with non-zero tokens.
-
-        Called from _run_sweep() every ~60 s. Also updates the cumulative
-        token counts on the open agent_spawn_sessions row so the DB reflects
-        current progress without waiting for session close.
-        Errors per-agent are caught so one bad agent doesn't abort the whole sweep.
-
-        Also publishes a USAGE_SNAPSHOT aggregate event after the loop so the
-        /ws/system dashboard updates live for active agents.
-        """
-        if not self._instances:
-            return
-
-        try:
-            from roboco.db.base import get_session_factory
-        except ImportError:
-            return
-
-        session_factory = get_session_factory()
-
-        # Accumulators for the post-loop USAGE_SNAPSHOT event.
-        _usage_by_agent: list[dict[str, Any]] = []
-        _usage_total_input = 0
-        _usage_total_output = 0
-        _usage_total_cost = 0.0
-
-        async with httpx.AsyncClient(timeout=3.0) as client:
-            for agent_id, instance in list(self._instances.items()):
-                if instance.state not in (
-                    AgentState.ACTIVE,
-                    AgentState.WAITING_SHORT,
-                ):
-                    continue
-
-                try:
-                    tokens = await self._resolve_active_tokens(client, agent_id)
-                    if tokens is None:
-                        continue
-
-                    persisted = await self._persist_token_snapshot(
-                        session_factory, agent_id, instance, tokens
-                    )
-                    if not persisted:
-                        continue
-
-                    tokens_input, tokens_output = tokens[0], tokens[1]
-                    model = instance.config.model if instance.config else "unknown"
-
-                    # Accumulate per-agent data for the aggregate snapshot.
-                    with contextlib.suppress(Exception):
-                        from roboco.billing.pricing import calculate_cost
-
-                        agent_cost = calculate_cost(
-                            model=model,
-                            tokens_input=tokens_input,
-                            tokens_output=tokens_output,
-                        )
-                        _usage_by_agent.append(
-                            {
-                                "agent_id": agent_id,
-                                "input_tokens": tokens_input,
-                                "output_tokens": tokens_output,
-                                "model": model,
-                                "cost_estimate": agent_cost,
-                            }
-                        )
-                        _usage_total_input += tokens_input
-                        _usage_total_output += tokens_output
-                        _usage_total_cost += agent_cost
-
-                except Exception as agent_exc:
-                    logger.debug(
-                        "Token snapshot failed for agent",
-                        agent_id=agent_id,
-                        error=str(agent_exc),
-                    )
-
-        # Publish a USAGE_SNAPSHOT aggregate if any active agents had token data.
-        if _usage_by_agent:
-            with contextlib.suppress(Exception):
-                from roboco.events import get_event_bus
-                from roboco.services.usage_events import (
-                    UsageSnapshot,
-                    publish_usage_snapshot,
-                )
-
-                await publish_usage_snapshot(
-                    get_event_bus(),
-                    UsageSnapshot(
-                        period="live",
-                        totals={
-                            "input_tokens": _usage_total_input,
-                            "output_tokens": _usage_total_output,
-                        },
-                        cost_estimate=_usage_total_cost,
-                        by_agent=_usage_by_agent,
-                    ),
-                )
-
-    async def _sweep_daily_rollup(self) -> None:
-        """Upsert daily_usage_rollups from closed agent_spawn_sessions.
-
-        Groups ended sessions by (date, agent_slug, team, model) and sums
-        their token counts + cost. Uses a Python-side upsert to stay
-        compatible with asyncpg / SQLAlchemy without raw INSERT ... ON CONFLICT
-        dialect-specific SQL.
-        Errors are caught so a bad rollup doesn't abort the sweeper.
-        """
-        try:
-            from roboco.db.base import get_session_factory
-            from roboco.db.tables import AgentSpawnSessionTable
-        except ImportError:
-            return
-
-        try:
-            from uuid import uuid4 as _uuid4
-
-            from sqlalchemy import func, select
-
-            session_factory = get_session_factory()
-            async with session_factory() as db:
-                # Aggregate closed sessions by (date, agent_slug, team, model).
-                # Limit to the last 7 days to avoid re-aggregating all-time
-                # history on every sweep — older days are already stable.
-                rollup_window_start = datetime.now(UTC) - timedelta(days=7)
-                result = await db.execute(
-                    select(
-                        func.date(AgentSpawnSessionTable.started_at).label("date"),
-                        AgentSpawnSessionTable.agent_slug,
-                        AgentSpawnSessionTable.team,
-                        AgentSpawnSessionTable.model,
-                        func.sum(AgentSpawnSessionTable.tokens_input).label(
-                            "tokens_input"
-                        ),
-                        func.sum(AgentSpawnSessionTable.tokens_output).label(
-                            "tokens_output"
-                        ),
-                        func.sum(AgentSpawnSessionTable.tokens_cache_read).label(
-                            "tokens_cache_read"
-                        ),
-                        func.sum(AgentSpawnSessionTable.tokens_cache_write).label(
-                            "tokens_cache_write"
-                        ),
-                        func.sum(AgentSpawnSessionTable.estimated_cost_usd).label(
-                            "total_cost_usd"
-                        ),
-                        func.count(AgentSpawnSessionTable.id).label("session_count"),
-                    )
-                    .where(
-                        AgentSpawnSessionTable.ended_at.isnot(None),
-                        AgentSpawnSessionTable.started_at >= rollup_window_start,
-                    )
-                    .group_by(
-                        func.date(AgentSpawnSessionTable.started_at),
-                        AgentSpawnSessionTable.agent_slug,
-                        AgentSpawnSessionTable.team,
-                        AgentSpawnSessionTable.model,
-                    )
-                )
-                rows = result.fetchall()
-
-                for row in rows:
-                    await self._upsert_rollup_row(db, row, _uuid4)
-
-                await db.commit()
-                logger.debug("Daily usage rollup complete", rows_processed=len(rows))
-
-        except Exception as exc:
-            logger.warning("Daily usage rollup failed", error=str(exc))
-
-    async def _upsert_rollup_row(self, db: Any, row: Any, uuid4: Any) -> None:
-        """Insert or update a single daily_usage_rollups row from an aggregate.
-
-        Looks up the existing rollup for (date, agent_slug, team, model) and
-        either updates its summed columns or inserts a fresh row.
-        """
-        from sqlalchemy import select, update
-
-        from roboco.db.tables import DailyUsageRollupTable
-
-        key = {
-            "date": row.date,
-            "agent_slug": row.agent_slug,
-            "team": row.team,
-            "model": row.model,
-        }
-        values = {
-            "tokens_input": int(row.tokens_input or 0),
-            "tokens_output": int(row.tokens_output or 0),
-            "tokens_cache_read": int(row.tokens_cache_read or 0),
-            "tokens_cache_write": int(row.tokens_cache_write or 0),
-            "total_cost_usd": float(row.total_cost_usd or 0.0),
-            "session_count": int(row.session_count or 0),
-        }
-
-        existing_result = await db.execute(
-            select(DailyUsageRollupTable).where(
-                DailyUsageRollupTable.date == key["date"],
-                DailyUsageRollupTable.agent_slug == key["agent_slug"],
-                DailyUsageRollupTable.team == key["team"],
-                DailyUsageRollupTable.model == key["model"],
-            )
-        )
-        existing = existing_result.scalar_one_or_none()
-
-        if existing is not None:
-            await db.execute(
-                update(DailyUsageRollupTable)
-                .where(DailyUsageRollupTable.id == existing.id)
-                .values(**values)
-            )
-        else:
-            db.add(DailyUsageRollupTable(id=uuid4(), **key, **values))
-
     async def restore_waiting_records(self) -> int:
         """Load persisted waiting records into memory on orchestrator start.
 
@@ -3448,286 +2903,6 @@ Start by:
 
     # =========================================================================
     # RATE-LIMIT PROBE LOOP
-    # =========================================================================
-
-    async def _rate_limit_probe_loop(self) -> None:
-        """Background loop: probe rate-limited providers every ~30 seconds.
-
-        Runs independently of the 60-second session/notification sweeper so
-        rate limits can be cleared on their own cadence without blocking
-        other sweep work.
-        """
-        probe_interval = 30  # seconds
-        while self._running:
-            try:
-                await asyncio.sleep(probe_interval)
-                await self._sweep_rate_limit_probes()
-            except asyncio.CancelledError:
-                break
-            except Exception as e:
-                logger.error("Rate-limit probe loop error", error=str(e))
-
-    async def _sweep_rate_limit_probes(self) -> None:
-        """One probe pass: check every rate-limited provider.
-
-        For each provider whose estimated_lift_at has passed:
-        - Call ``_do_probe(provider)`` to test connectivity.
-        - **Success**: clear the tracker, resolve all parked agents, publish
-          ``RATE_LIMIT_LIFTED``.
-        - **Failure**: increment probe_failures; if the count reaches 10 and
-          we haven't already sent a CEO notification for this episode, send
-          one now.
-        """
-        from roboco.services.gateway.rate_limit_tracker import RateLimitStateTracker
-
-        try:
-            providers = await RateLimitStateTracker.list_rate_limited_providers()
-        except Exception as e:
-            logger.warning("Failed to list rate-limited providers", error=str(e))
-            return
-
-        for provider, state in providers:
-            try:
-                await self._probe_one_provider(provider, state)
-            except Exception as e:
-                logger.error(
-                    "Unhandled error probing provider",
-                    provider=provider,
-                    error=str(e),
-                )
-
-    def _make_tracker(self, provider: str) -> Any:
-        """Return a RateLimitStateTracker for *provider*.
-
-        Extracted as its own method so unit tests can monkeypatch it to
-        return an async mock without needing to intercept lazy imports.
-        """
-        from roboco.services.gateway.rate_limit_tracker import RateLimitStateTracker
-
-        return RateLimitStateTracker(provider)
-
-    @staticmethod
-    def _too_early_to_probe(state: dict[str, Any]) -> bool:
-        """True while the estimated lift time (activated_at + retry_after) is future.
-
-        Missing or malformed timestamps fall through to allow the probe.
-        """
-        activated_at_raw = state.get("activated_at")
-        retry_after = state.get("retry_after")
-        if not activated_at_raw or retry_after is None:
-            return False
-        try:
-            activated_at = datetime.fromisoformat(activated_at_raw)
-        except (ValueError, TypeError):
-            return False
-        return datetime.now(UTC) < activated_at + timedelta(seconds=retry_after)
-
-    def _parked_agents_for(self, provider: str) -> list[str]:
-        """Agent slugs parked waiting for *provider*'s rate limit to lift."""
-        return [
-            agent_id
-            for agent_id, record in list(self._waiting_records.items())
-            if record.waiting_for == "rate_limit_lifted"
-            and record.context.get("provider") == provider
-        ]
-
-    async def _on_probe_success(self, provider: str, tracker: Any) -> None:
-        """Clear the limit, resume parked agents, publish RATE_LIMIT_LIFTED."""
-        logger.info("Rate-limit probe succeeded; clearing provider", provider=provider)
-        await tracker.clear()
-        # New episodes should get a fresh CEO notification.
-        self._rate_limit_ceo_notified.discard(provider)
-        resumed = self._parked_agents_for(provider)
-        for agent_id in resumed:
-            with contextlib.suppress(Exception):
-                await self.resolve_wait(
-                    agent_id,
-                    {
-                        "reason": "rate_limit_lifted",
-                        "provider": provider,
-                        "lifted_at": datetime.now(UTC).isoformat(),
-                    },
-                )
-        with contextlib.suppress(Exception):
-            from roboco.events import get_event_bus
-            from roboco.models.events import Event, EventType
-
-            await get_event_bus().publish(
-                Event(
-                    type=EventType.RATE_LIMIT_LIFTED,
-                    data={
-                        "provider": provider,
-                        "resumedAgents": resumed,
-                        "timestamp": datetime.now(UTC).isoformat(),
-                    },
-                )
-            )
-        logger.info(
-            "RATE_LIMIT_LIFTED published",
-            provider=provider,
-            resumed_agents=len(resumed),
-        )
-
-    async def _on_probe_failure(
-        self, provider: str, tracker: Any, activated_at_raw: str | None
-    ) -> None:
-        """Count a failed probe; notify the CEO once at the failure threshold."""
-        failure_count = await tracker.increment_probe_failures()
-        logger.debug(
-            "Rate-limit probe failed", provider=provider, probe_failures=failure_count
-        )
-        if (
-            failure_count >= _CEO_NOTIFY_THRESHOLD
-            and provider not in self._rate_limit_ceo_notified
-        ):
-            self._rate_limit_ceo_notified.add(provider)
-            await self._notify_rate_limit_ceo(
-                provider=provider,
-                activated_at_str=activated_at_raw or "unknown",
-                paused_agent_count=len(self._parked_agents_for(provider)),
-            )
-
-    async def _probe_one_provider(self, provider: str, state: dict[str, Any]) -> None:
-        """Probe a single rate-limited provider and handle the outcome."""
-        if self._too_early_to_probe(state):
-            return  # Wait until after the estimated lift time.
-        tracker = self._make_tracker(provider)
-        if await self._do_probe(provider):
-            await self._on_probe_success(provider, tracker)
-        else:
-            await self._on_probe_failure(provider, tracker, state.get("activated_at"))
-
-    @staticmethod
-    def _probe_target(provider: str) -> tuple[str | None, dict[str, str]]:
-        """Resolve the (url, headers) for a free liveness probe of ``provider``.
-
-        Returns ``(None, {})`` when the provider can't be probed — an unknown
-        provider, or Anthropic with no API key configured. The caller then
-        falls back to time-expiry optimism rather than parking forever.
-        """
-        p = provider.lower()
-        if p == "anthropic":
-            key = settings.anthropic_api_key
-            if not key:
-                return None, {}
-            return (
-                f"{_ANTHROPIC_PROBE_BASE}/v1/models",
-                {"x-api-key": key, "anthropic-version": "2023-06-01"},
-            )
-        if p.startswith("ollama"):
-            return f"{settings.ollama_base_url.rstrip('/')}/api/tags", {}
-        return None, {}
-
-    async def _do_probe(self, provider: str) -> bool:
-        """Return True if ``provider`` is accepting requests again (not 429).
-
-        Makes a free, unmetered liveness call — Anthropic ``GET /v1/models``
-        or Ollama ``GET /api/tags`` — and treats any non-429 response as the
-        rate limit having lifted. A 429 keeps the provider parked; a network
-        error stays parked too (retry next sweep). When the provider can't be
-        probed (no key / unknown), fall back to time-expiry optimism: the
-        caller only reaches this after ``estimated_lift_at`` has passed.
-
-        Injectable boundary — tests monkeypatch this to force outcomes.
-        """
-        url, headers = self._probe_target(provider)
-        if url is None:
-            return True  # cannot probe — trust the elapsed retry_after window
-        try:
-            async with httpx.AsyncClient(timeout=_PROBE_TIMEOUT_SECONDS) as client:
-                resp = await client.get(url, headers=headers)
-        except httpx.HTTPError as exc:
-            logger.debug(
-                "Rate-limit probe request failed", provider=provider, error=str(exc)
-            )
-            return False  # unreachable — stay parked, retry on the next sweep
-        return resp.status_code != _HTTP_TOO_MANY_REQUESTS
-
-    async def _notify_rate_limit_ceo(
-        self,
-        provider: str,
-        activated_at_str: str,
-        paused_agent_count: int,
-    ) -> None:
-        """Send a high-priority notification to the CEO about a persistent rate limit.
-
-        Fires once per rate-limit episode. Follows the same pattern as
-        ``_notify_stranded_agent`` — direct DB insert + delivery.deliver().
-        """
-        try:
-            from sqlalchemy import select as _select
-
-            from roboco.db.base import get_session_factory
-            from roboco.db.tables import AgentTable, NotificationTable
-            from roboco.models.base import (
-                AgentRole,
-                NotificationPriority,
-                NotificationType,
-            )
-            from roboco.services.notification_delivery import (
-                get_notification_delivery_service,
-            )
-            from roboco.utils.converters import require_uuid
-
-            # Compute human-friendly duration
-            duration_desc = "unknown duration"
-            try:
-                activated_at = datetime.fromisoformat(activated_at_str)
-                elapsed = datetime.now(UTC) - activated_at
-                total_minutes = int(elapsed.total_seconds() / 60)
-                if total_minutes < 60:  # noqa: PLR2004
-                    duration_desc = f"{total_minutes} minute(s)"
-                else:
-                    duration_desc = f"{total_minutes // 60}h {total_minutes % 60}m"
-            except (ValueError, TypeError):
-                pass
-
-            session_factory = get_session_factory()
-            async with session_factory() as db:
-                ceo_result = await db.execute(
-                    _select(AgentTable).where(AgentTable.role == AgentRole.CEO)
-                )
-                ceo = ceo_result.scalar_one_or_none()
-                if ceo is None:
-                    logger.warning(
-                        "CEO agent not found; skipping rate-limit CEO notification",
-                        provider=provider,
-                    )
-                    return
-                notification = NotificationTable(
-                    type=NotificationType.ALERT,
-                    priority=NotificationPriority.HIGH,
-                    from_agent=ceo.id,
-                    to_agents=[ceo.id],
-                    subject=f"Rate limit persisting: {provider}",
-                    body=(
-                        f"Provider '{provider}' has been rate-limited for "
-                        f"{duration_desc}. "
-                        f"{paused_agent_count} agent(s) are currently paused. "
-                        f"10 consecutive probe attempts have failed. "
-                        f"Manual intervention may be required."
-                    ),
-                    requires_ack=True,
-                )
-                db.add(notification)
-                await db.flush()
-                delivery = get_notification_delivery_service(db)
-                await delivery.deliver(require_uuid(notification.id))
-                await db.commit()
-            logger.info(
-                "Rate-limit CEO notification sent",
-                provider=provider,
-                paused_agents=paused_agent_count,
-            )
-        except Exception as e:
-            logger.error(
-                "Failed to send rate-limit CEO notification",
-                provider=provider,
-                error=str(e),
-            )
-
-    # =========================================================================
-    # STATUS API
     # =========================================================================
 
     def get_state(self, agent_id: str) -> AgentState:
@@ -6681,354 +5856,3 @@ Run the project's quality checks against acceptance criteria:
             state, f'Call evidence(task_id="{task_id}") to check status.'
         )
 
-    def _build_dev_prompt(self, task: dict[str, Any]) -> str:
-        """Build state-aware initial prompt for a developer."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        status = task.get("status", "unknown")
-
-        # Determine workflow state based on task attributes
-        has_plan = bool(task.get("plan"))
-        workflow_state = self._get_workflow_state(status, has_plan)
-        instructions = self._get_workflow_instructions(workflow_state, task_id)
-
-        return f"""You have been assigned a development task.
-
-TASK ID: {task_id}
-TITLE: {title}
-STATUS: {status}
-WORKFLOW STATE: {workflow_state}
-
-{instructions}
-
-Start by calling evidence(task_id="{task_id}") for full details and acceptance criteria.
-
-When out of work: i_am_idle().
-"""
-
-    def _build_qa_prompt(self, task: dict[str, Any]) -> str:
-        """Build initial prompt for a QA agent."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        assigned_to = task.get("assigned_to", "unknown")
-        team = task.get("team", "unknown")
-
-        return f"""A task is ready for QA review.
-
-TASK ID: {task_id}
-TITLE: {title}
-DEVELOPER: {assigned_to}
-TEAM: {team}
-
-== QA WORKFLOW ==
-
-1. claim_review(task_id="{task_id}")
-   — assigns the QA seat; returns inline diff + PR + commits as evidence.
-   The PR is already open (dev opened it before submitting QA);
-   review on GitHub if you need more context.
-2. Review the implementation against EVERY acceptance criterion.
-   Run/read tests; sanity-check the diff for regressions, security,
-   and scope creep.
-3. Decide:
-   - PASS: pass(task_id="{task_id}",
-            notes="<>=80 chars: what you verified, which AC, evidence>")
-     — transitions awaiting_qa → awaiting_documentation.
-   - FAIL: fail(task_id="{task_id}",
-            issues=["concrete issue 1", "concrete issue 2", ...])
-     — transitions to needs_revision; each issue must be specific and
-     actionable.
-4. note(scope='reflect'|'learning', task_id="{task_id}", text=...)
-   for anything worth flagging.
-5. give_me_work() to pick up the next QA item,
-   or i_am_idle() if the queue is empty.
-"""
-
-    def _build_doc_prompt(self, task: dict[str, Any]) -> str:
-        """Build initial prompt for a documenter."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        team = task.get("team", "unknown")
-
-        return f"""A task is ready for documentation. The dev's PR is already open
-— you're documenting alongside the QA-passed branch.
-
-TASK ID: {task_id}
-TITLE: {title}
-TEAM: {team}
-
-== DOC WORKFLOW ==
-
-1. claim_doc_task(task_id="{task_id}")
-   — assigns the doc seat and opens your workspace on the task's branch.
-2. evidence(task_id="{task_id}") — read dev handoff notes, qa_notes,
-   and the inline diff so the docs reflect what actually shipped.
-3. Write/update docs in your workspace: README sections, API references,
-   code comments, migration notes, or new docs files as the change requires.
-4. commit("docs(scope): <subject, >=20 chars>") per logical doc chunk
-   — auto-prefixes the task ID and stages tracked changes.
-5. i_documented(task_id="{task_id}",
-   notes="<>=20 chars: what you documented and where>",
-   files=["docs/foo.md", "README.md", ...])
-   — transitions awaiting_documentation → awaiting_pm_review.
-6. give_me_work() for the next doc item,
-   or i_am_idle() if the queue is empty.
-"""
-
-    def _build_pm_review_prompt(self, task: dict[str, Any]) -> str:
-        """Prompt for PM reviewing a SUBTASK in awaiting_pm_review."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        team = task.get("team", "unknown")
-
-        return f"""A SUBTASK in your cell is awaiting your PM review.
-It has passed QA and documentation; the leaf PR is open and ready to merge.
-
-TASK ID: {task_id}
-TITLE: {title}
-TEAM: {team}
-
-== PM REVIEW WORKFLOW (leaf subtask) ==
-
-1. evidence(task_id="{task_id}")
-   — review PR, commits, inline diff, dev_notes, qa_notes, doc files.
-2. Spot-check that:
-   - every acceptance criterion is satisfied,
-   - QA's pass notes line up with the actual diff,
-   - docs reflect what shipped.
-3. note(scope='decision', task_id="{task_id}",
-        text="<approve rationale or rejection reason>")
-   — REQUIRED before complete().
-4. Decide:
-   - APPROVE: complete(task_id="{task_id}", notes="<merge rationale>")
-     — auto-merges the leaf PR and finalizes the subtask.
-   - NEEDS REWORK: leave a clear note(scope='decision', text="...") and
-     rely on the dispatcher to respawn the dev for revision.
-     Use escalate_up only if the issue is truly outside your cell.
-5. give_me_work() / triage() for the next item, or i_am_idle().
-
-Never `commit`, never write code, never run `git`. PMs coordinate.
-"""
-
-    def _build_board_prompt(self, task: dict[str, Any]) -> str:
-        """Prompt for a board agent (Product Owner / Head of Marketing) to
-        review and SHAPE a strategic task. Board roles advise — they do not
-        build, code, or delegate."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        description = task.get("description", "No description")
-
-        return f"""\
-You are on the Board. This strategic task is under board review.
-
-TASK: {task_id}
-TITLE: {title}
-DESCRIPTION: {description}
-
-THE BOARD REVIEWS AS A PAIR: the Product Owner AND the Head of Marketing both
-review every board task before it reaches the CEO. The Product Owner owns
-product requirements + acceptance scope; the Head of Marketing owns the UX /
-user-facing / positioning dimension. The CEO only gets the handoff after BOTH
-of you have recorded a review.
-
-YOUR ROLE: review and shape this work. You do NOT build, code, claim, or
-delegate — those verbs are not yours. Your deliverable is a recorded review.
-
-== WHAT TO DO ==
-
-1. triage()
-     — see your board-level work and context.
-2. note(text="<the product requirements and acceptance criteria you expect, the
-        scope, the must-haves, and what 'done' looks like — Head of Marketing:
-        the UX, user-facing impact, and how the feature is positioned>",
-        scope='decision', task_id="{task_id}")
-     — this recorded review is how the CEO and Main PM act on your input.
-3. say(...) in your board channel to flag UX, positioning, or risk concerns and
-     to coordinate with your fellow board reviewer.
-4. i_am_idle()
-     — when your review is recorded. Once both board reviewers are done, the
-       CEO is notified the task is ready for Approve & Start, then routes it to
-       Main PM for delegation to the cells; you do NOT hand it off yourself.
-
-Do NOT attempt to claim, plan, complete, or delegate — the gateway will reject
-those, and a substantive recorded note IS your job here.
-"""
-
-    def _build_marketing_prompt(self, task: dict[str, Any]) -> str:
-        """Build initial prompt for head-marketing with a marketing task."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        description = task.get("description", "No description")
-
-        return f"""You have been assigned a marketing task.
-
-TASK ID: {task_id}
-TITLE: {title}
-DESCRIPTION: {description}
-
-Begin work:
-
-1. Review the task details above (full acceptance criteria arrive in your
-   briefing / the give_me_work response)
-2. Execute the marketing task (content, campaigns, research, etc.)
-3. Coordinate with Product Owner or Main PM if needed
-4. Call i_am_done() when done
-5. Call give_me_work() to check for more marketing work
-6. If no more work, call i_am_idle() to shutdown gracefully
-"""
-
-    def _build_pm_blocker_prompt(self, task: dict[str, Any]) -> str:
-        """Build initial prompt for a Cell PM handling a blocker."""
-        task_id = task.get("id", "unknown")
-        title = task.get("title", "Untitled")
-        assigned_to = task.get("assigned_to", "unknown")
-        blocker = task.get("blocker", {})
-        reason = blocker.get("reason", "Unknown")
-        what_needed = blocker.get("what_needed", "Unknown")
-
-        return f"""A task in your cell is BLOCKED and needs your attention.
-
-TASK ID: {task_id}
-TITLE: {title}
-ASSIGNED TO: {assigned_to}
-BLOCKER REASON: {reason}
-WHAT'S NEEDED: {what_needed}
-
-Your job:
-
-1. Understand the blocker by reviewing task details
-2. Communicate with the blocked developer if needed
-3. Resolve the blocker (coordinate resources, make decisions, escalate if needed)
-4. Once resolved, call unblock("{task_id}") to release the task back to the developer
-5. Call triage() to check for other blocked tasks in your cell
-6. If no more blockers, call i_am_idle() to shutdown gracefully
-"""
-
-    def _build_escalation_prompt(self, notification: dict[str, Any]) -> str:
-        """Build initial prompt for handling an escalation."""
-        notif_id = notification.get("id", "unknown")
-        from_agent = notification.get("from_agent", "unknown")
-        subject = notification.get("subject", "No subject")
-        priority = notification.get("priority", "normal")
-        body = notification.get("body", "No details provided")
-
-        return f"""You have received an ESCALATION that requires your attention.
-
-FROM: {from_agent}
-SUBJECT: {subject}
-PRIORITY: {priority}
-
-DETAILS:
-{body}
-
-Your job:
-
-1. Acknowledge the notification with notify_ack("{notif_id}")
-2. Assess the escalation and determine action needed
-3. Communicate decisions via appropriate channels
-4. If this requires further escalation, use escalate_up()
-5. When resolved, call triage() for other work
-6. If no more work, call i_am_idle() to shutdown gracefully
-"""
-
-    def _build_approval_prompt(self, notification: dict[str, Any]) -> str:
-        """Build initial prompt for handling an approval request."""
-        notif_id = notification.get("id", "unknown")
-        from_agent = notification.get("from_agent", "unknown")
-        subject = notification.get("subject", "No subject")
-        related_task_id = notification.get("related_task_id", "None")
-        body = notification.get("body", "No details provided")
-
-        return f"""You have received an APPROVAL REQUEST.
-
-FROM: {from_agent}
-SUBJECT: {subject}
-RELATED TASK: {related_task_id}
-
-REQUEST:
-{body}
-
-Your job:
-
-1. Review the approval request carefully
-2. If related to a task, use the task context provided in your briefing
-3. Make your decision and communicate it
-4. Acknowledge with notify_ack("{notif_id}")
-5. Call triage() for other work
-6. If no more work, call i_am_idle() to shutdown gracefully
-"""
-
-    def _build_audit_prompt(self, alert: dict[str, Any] | None = None) -> str:
-        """Build initial prompt for the auditor."""
-        if alert:
-            subject = alert.get("subject", "Quality issue detected")
-            body = alert.get("body", "Review system quality metrics")
-
-            return f"""QUALITY ALERT triggered your attention.
-
-ALERT: {subject}
-DETAILS: {body}
-
-Your job:
-
-1. Investigate the quality issue
-2. Review relevant channels and task history (you have read access to all)
-3. Compile your findings
-4. Report to CEO via appropriate channel
-5. Call i_am_idle() when complete
-"""
-
-        return """Periodic AUDIT requested.
-
-Your job:
-
-1. Review recent activity across all cells
-2. Check quality metrics (QA pass/fail rates, blocker frequency, etc.)
-3. Identify any concerns or patterns
-4. Compile audit report for CEO
-5. Call i_am_idle() when complete
-"""
-
-    def _build_a2a_prompt(self, notification: dict[str, Any]) -> str:
-        """Build initial prompt for handling an A2A (Agent-to-Agent) request.
-
-        Reads `priority` directly off the notification row (set by
-        NotificationService.send_a2a_notification). Pre-Phase-3 this
-        consumed a non-existent `metadata.urgent` and always rendered
-        urgency_note=False; the column-level priority is now the source
-        of truth.
-        """
-        notif_id = notification.get("id", "unknown")
-        from_agent = notification.get("from_agent", "unknown")
-        body = notification.get("body", "No message provided")
-        related_task_id = notification.get("related_task_id")
-        metadata = notification.get("metadata", {})
-        skill = metadata.get("skill", "general")
-        priority_raw = notification.get("priority", "normal")
-
-        # URGENT gets the bold attention-grabber; HIGH gets a quieter
-        # "higher priority" hint; NORMAL gets no prefix.
-        if priority_raw == "urgent":
-            urgency_note = "**URGENT** - This request has priority.\n\n"
-        elif priority_raw == "high":
-            urgency_note = "**HIGH PRIORITY** - Please handle promptly.\n\n"
-        else:
-            urgency_note = ""
-        task_note = f"RELATED TASK: {related_task_id}\n" if related_task_id else ""
-
-        return f"""You have received an A2A (Agent-to-Agent) REQUEST.
-
-{urgency_note}FROM: {from_agent}
-SKILL: {skill}
-{task_note}
-REQUEST:
-{body}
-
-Your job:
-
-1. Acknowledge the notification with notify_ack("{notif_id}")
-2. Process the request using your {skill} capabilities
-3. Respond to {from_agent} using dm("{from_agent}", ...)
-4. If you need task context, it is provided in your briefing for the related task
-5. When done, call give_me_work() for other work
-6. If no more work, call i_am_idle() to shutdown gracefully
-"""
